+++ src/datasources/work-orders/components/WorkOrdersQueryEditor.test.tsx
@@ -5,7 +5,7 @@ import { OutputType, WorkOrderProperties, WorkOrderPropertiesOptions, WorkOrders
 import { QueryEditorProps } from '@grafana/data';
 import React from 'react';
 import userEvent from '@testing-library/user-event';
-import { select } from 'react-select-event';
+import selectEvent, { select } from 'react-select-event';
 
 const mockOnChange = jest.fn();
 const mockOnRunQuery = jest.fn();
@@ -64,6 +64,13 @@ describe('WorkOrdersQueryEditor', () => {
     const descending = container.getByRole('checkbox');
     expect(descending).toBeInTheDocument();
     expect(descending).not.toBeChecked();
+
+    expect(mockOnChange).toHaveBeenCalledWith(
+      expect.objectContaining({
+        outputType: OutputType.Properties,
+        refId: 'A'
+      }));
+    expect(mockOnRunQuery).toHaveBeenCalledTimes(1);
   });
 
   describe('output type is total count', () => {
@@ -132,6 +139,13 @@ describe('WorkOrdersQueryEditor', () => {
     it('should render order by', async () => {
       const orderBy = container.getAllByRole('combobox')[1];
       expect(orderBy).toBeInTheDocument();
+
+     selectEvent.openMenu(orderBy);
+
+     expect(screen.getByText('ID')).toBeInTheDocument();
+     expect(screen.getByText('ID of the work order')).toBeInTheDocument();
+     expect(screen.getByText('Updated At')).toBeInTheDocument();
+     expect(screen.getByText('Latest update at time of the work order')).toBeInTheDocument();
     });
 
     it('should render descending', async () => {
@@ -210,6 +224,44 @@ describe('WorkOrdersQueryEditor', () => {
       await waitFor(() => {
         expect(mockOnChange).toHaveBeenCalledWith(expect.objectContaining({ descending: true }));
         expect(mockOnRunQuery).toHaveBeenCalled();
+      });
+    });
+
+    it('should call onChange when query by changes', async () => {
+      const container = renderElement();
+
+      const queryBuilder = container.getByRole('dialog');
+      expect(queryBuilder).toBeInTheDocument();
+
+      // Simulate a change event
+      const event = { detail: { linq: 'new-query' } };
+      queryBuilder?.dispatchEvent(new CustomEvent('change', event));
+
+      await waitFor(() => {
+        expect(mockOnChange).toHaveBeenCalledWith(expect.objectContaining({ queryBy: 'new-query' }));
+        expect(mockOnRunQuery).toHaveBeenCalled();
+      });
+    });
+
+    it('should not call onChange when query by changes with same value', async () => {
+      const container = renderElement();
+      mockOnChange.mockClear();
+      mockOnRunQuery.mockClear();
+
+      const queryBuilder = container.getByRole('dialog');
+      expect(queryBuilder).toBeInTheDocument();
+
+      // Simulate a change event
+      let event = { detail: { linq: 'new-query' } };
+      queryBuilder?.dispatchEvent(new CustomEvent('change', event));
+
+      // Simulate a change event with the same value
+      event = { detail: { linq: 'new-query' } };
+      queryBuilder?.dispatchEvent(new CustomEvent('change', event));
+
+      await waitFor(() => {
+        expect(mockOnChange).toHaveBeenCalledTimes(1);
+        expect(mockOnRunQuery).toHaveBeenCalledTimes(1);
       });
     });
   });